pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-healthcare-intel"
version = "0.1.0"
description = "Agentic Healthcare Intelligence System for Indian medical facility discovery and trust scoring"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "pyarrow>=16.1.0",
  "fastapi>=0.112.0",
  "uvicorn>=0.30.0",
  "streamlit>=1.37.0",
  "plotly>=5.22.0",
  "mlflow>=2.15.0",
  "databricks-vectorsearch>=0.38.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
